requirements.txt
mysql-connector-python
nicegui
pyperclip